Move PartyListHUDView
